requirements.txt
anyio==4.6.2.post1
blinker==1.9.0
certifi==2024.8.30
click==8.1.7
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dotenv==1.0.1
python-telegram-bot==21.8
sniffio==1.3.1
Werkzeug==3.1.3
Flask-SQLAlchemy==3.1.1